requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
executing==1.2.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
humanfriendly==10.0
idna==3.4
imageio==2.25.0
imgviz==1.6.2
importlib-metadata==6.0.0
ipython==8.9.0
jedi==0.18.2
keras==2.11.0
kiwisolver==1.4.4
labelme==5.1.1
lap==0.4.0
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.2.1
natsort==8.2.0
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
onnx==1.13.0
onnxruntime==1.13.1
onnxsim==0.4.13
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
prompt-toolkit==3.0.36
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
Pygments==2.14.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
setuptools==65.6.3
shiboken6==6.4.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tifffile==2023.1.23.1
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.12.0
